+++ crates/compiler/src/export.rs
@@ -6,9 +6,11 @@
 
 pub mod ca65;
 pub mod pvsneslib;
+pub mod tass64;
 
 pub use self::ca65::{Ca65Exporter, Ca65MemoryMap};
 pub use self::pvsneslib::{PvExporter, PvMemoryMap};
+pub use self::tass64::{Tass64Exporter, Tass64MemoryMap};
 
 use crate::audio_driver;
 use crate::common_audio_data::CommonAudioData;